### 2019-12-12 #1946 @kyleniemeyer
@openjournals/dev any ideas why the paper won't compile? The error message is not too helpful for me.
