### 2019-12-12 #1946 @danielskatz
could be a problem with a file - is the bib file correctly named and does it exist?  or any images?
