### 2019-12-12 #1946 @kyleniemeyer
The bib file looks ok, but there is an image in the article: `![Figure 1](JOSS_figure_workflow.png)`

I do see a `JOSS_figure_workflow.png` in the repo, in the same location as the `paper.md`, so I'm not sure why that would cause an error.
