### 2019-12-13 #1946 @arfon
A couple of issues here:

- Whedon is likely timing out here as the repository is _very_ large (2.3GB!) even though there's only ~104MB of files present in the repository.
- Also, the bibtex has errors that I've fixed in https://github.com/rgmyr/corebreakout/pull/3

Finally, here's a copy of the paper I compiled locally: [10.21105.joss.01946.pdf](https://github.com/openjournals/joss-reviews/files/3958474/10.21105.joss.01946.pdf)
