### 2019-12-13 #1946 @kyleniemeyer
@arfon  @rgmyr wow! That is a huge repository... is the large size due to files that were removed but are in the history?

If you know the files, you can remove them from the history using this: https://help.github.com/en/github/authenticating-to-github/removing-sensitive-data-from-a-repository#purging-a-file-from-your-repositorys-history

This is likely going to continue to be an issue, not only for whedon but for others looking to clone the repo, and so worth addressing.
