### 2019-12-13 #1946 @rgmyr
Sounds good, thanks for the help guys. I wasn't sure what the problem was with paper compilation -- it seemed to be working via the preview service for awhile, and then it just stopped being able to compile all of the sudden.

I've merged @arfon's PR, and will look through the history to see where that extra size is coming from... I'm not sure of the top of my head, but I wouldn't be surprised if there were a few model files or data folders hiding in there accidentally.
