+++ paper.bib
@@ -1,0 +1,87 @@
+@article{He:2017,
+  author    = {Kaiming He and
+               Georgia Gkioxari and
+               Piotr Doll{\'{a}}r and
+               Ross B. Girshick},
+  title     = {Mask R-CNN},
+  journal   = {CoRR},
+  volume    = {abs/1703.06870},
+  year      = {2017},
+  url       = {http://arxiv.org/abs/1703.06870},
+  archivePrefix = {arXiv},
+  eprint    = {1703.06870},
+  timestamp = {Mon, 13 Aug 2018 16:46:36 +0200},
+  biburl    = {https://dblp.org/rec/bib/journals/corr/HeGDG17},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@article{Russell:2007,
+    title={LabelMe: A Database and Web-Based Tool for Image Annotation},
+    volume={77},
+    DOI={10.1007/s11263-007-0090-8},
+    number={1-3},
+    journal={International Journal of Computer Vision},
+    author={Russell, Bryan C. and Torralba, Antonio and Murphy, Kevin P. and Freeman, William T.},
+    year={2007},
+    pages={157–173}
+}
+
+@article{Martin:2019,
+  author = {Thomas Martin and
+            Ross Meyer and
+            Zane Jobe},
+  title = {Lithology Prediction of Slabbed Core Photos Using Machine Learning Models},
+  year = {2019},
+  month = {4},
+  url = {https://figshare.com/articles/Lithology_Prediction_of_Slabbed_Core_Photos_Using_Machine_Learning_Models/8023835},
+  doi = {10.6084/m9.figshare.8023835.v2}
+}
+
+@article{Jobe:2017,
+    author = {Jobe, Zane and
+          Sylvester, Zoltán and
+          Pittaluga, Michele Bolla and
+          Frascati, Alessandro and
+          Pirmez, Carlos and
+          Minisini, Daniel and
+          Howes, Nick and
+          Cantelli, Alessandro},
+    title = {Facies architecture of submarine channel deposits on the western Niger Delta slope: Implications for grain-size and density stratification in turbidity currents},
+    journal = {Journal of Geophysical Research: Earth Surface},
+    volume = {122},
+    number = {2},
+    pages = {473-491},
+    keywords = {turbidity current, submarine channel, grain size, bed thickness, sediment concentration, channel margin},
+    doi = {10.1002/2016JF003903},
+    url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1002/2016JF003903},
+    eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1002/2016JF003903},
+    year = {2017}
+}
+
+@misc{Abdulla:2017,
+  title={Mask R-CNN for object detection and instance segmentation on Keras and TensorFlow},
+  author={Waleed Abdulla},
+  year={2017},
+  publisher={Github},
+  journal={GitHub repository},
+  howpublished={\url{https://github.com/matterport/Mask_RCNN}},
+}
+
+@misc{Wada:2016,
+  author =       {Ketaro Wada},
+  title =        {{labelme: Image Polygonal Annotation with Python}},
+  howpublished = {\url{https://github.com/wkentaro/labelme}},
+  year =         {2016}
+}
+
+@inproceedings{Smith:2007,
+  author = {Ray Smith},
+  title = {An Overview of the Tesseract OCR Engine},
+  booktitle = {ICDAR '07: Proceedings of the Ninth International Conference on Document Analysis and Recognition},
+  url = {http://www.google.de/research/pubs/archive/33418.pdf},
+  year = {2007},
[…]
