### 2019-12-13 25ec717
remove extra contrib file
Former-commit-id: fac75b4b98e1b8c010e5ab839c9f3113e77cfcad
Former-commit-id: c4cc5cddd4744f9edbc5bf54f98f611926dad375
