+++ CONTRIBUTING_CODE.md
@@ -1,8 +1,0 @@
-# Contributing Code
-
-1. Fork it!
-2. Create your feature branch
-3. Pass all `pytest` tests and verify `tests/notebooks/` run without errors
-4. Update `CHANGELOG.md`
-    - Describe changes under `[Unreleased]` section
-5. Submit a pull request
