### 2019-12-13 #1946 @rgmyr
I was having a hard time tracking down and removing the offending files/folders individually, so I tried a couple of things:
- Created a new `--orphan` branch, set that as `master`, and removed some old temporary branches that aren't relevant anymore. (Following [this stackoverflow answer](https://stackoverflow.com/questions/9683279/make-the-current-commit-the-only-initial-commit-in-a-git-repository/13102849#13102849).)
- That didn't seem to actually change the size much, if at all, so I also tried cleaning things up [BFG Repo-Cleaner](https://rtyley.github.io/bfg-repo-cleaner/).

I ended up having to be pretty aggressive with the latter. I think a lot of files were deleted from the history, but I made a local backup incase I need to restore for some reason, and I don't foresee it being a big problem for code management going forward.

In any case, the repo size is now 182MB, and the paper compiles pretty quickly on the preview service.
