### 2019-12-13 7434cd4
fix paper ref and figure text
