### 2019-12-13 #1946 @rgmyr
Just made a couple of small fixes (multiple reference separator, figure text)...
