### 2019-12-14 176d0aa
Update paper.md
Just adding a middle initial
