### 2019-12-14 #1946 @ThomasMGeo
Maybe @brendonhall or @kwinkunks would be interested in reviewing?
