### 2019-12-15 #1946 @rgmyr
Hi @brendonhall, here's the [Reviewer Guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Once a [Review] issue is opened, each reviewer gets a [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) to work through (check out some of the review issues in this repo for examples).
