### 2019-12-16 #1946 @kbarnhart
@rgmyr thank you for the recommendations. I'm now fully back online after a week at AGU.

@brendonhall thank you for agreeing to review this publication. In my next comment, I'll formally add you.

:wave: @kwinkunks would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is a pre-review issue that I use to coordinate reviewers. Once there are sufficient reviewers (usually 2-3), I will open a new issue where the review will take place.

If you are not able, please let me know if you have recommendations for other reviewers.
