### 2019-12-17 #1946 @JesperDramsch
This seems awesome. I'd love to take a look and review this.

I'll probably be best on the python and ml bits and ok on the core.
