### 2019-12-17 #1946 @ThomasMGeo
Luckily it's way more the python/ML bits, and just being able to say rock vs no rock! Thanks @JesperDramsch
