### 2019-12-17 #1946 @kbarnhart
Thanks for being willing to review @JesperDramsch

Thanks also @ThomasMGeo for your assistance in finding appropriate reviewers.

:wave: @swgeoscience @jessepisel can you confirm on this thread that would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is a pre-review issue that I use to coordinate reviewers. Once there are sufficient reviewers (usually 2-3, it looks like this submission may have 4), I will open a new issue where the review will take place.
