### 2019-12-18 #1969 @whedon
[REVIEW]: CoreBreakout: Subsurface Core Images to Depth-Registered Datasets
**Submitting author:** @rgmyr (<a href="http://orcid.org/0000-0003-2344-554X">Ross Meyer</a>)
**Repository:** <a href="https://github.com/rgmyr/corebreakout" target ="_blank">https://github.com/rgmyr/corebreakout</a>
**Version:** v0.3
