### 2019-12-18 #1946 @kbarnhart
Thanks all for your enthusiasm for finding reviewers and recommending reviewers. I've started the formal review over at #1969. @swgeoscience, if you are really excited to also be a reviewer, let me know and I can add you over on that issue.

For those of you who are interested in being listed on our [list of potential reviewers](https://bit.ly/joss-reviewers) for future submissions, please consider filling out the 1-minute form at [this link](https://joss.theoj.org/reviewer-signup.html). This helps editors and authors find appropriate reviewers.
