### 2019-12-20 #1969 @JesperDramsch
👋🏼 @kbarnhart During the pre-review @rgmyr was asked to reduce the repo size (from 2+GB), which was accomplished by creating a new orphan. This, however, removed Zane's contributions completely from the history and Thomas' contribution to the update of the `paper.md`. [See comment in #1946](https://github.com/openjournals/joss-reviews/issues/1946#issuecomment-565326410).

From @arfon's [fork (pre-orphan)](https://github.com/arfon/corebreakout) with a rich editing history, it seems like the contribution and authorship criteria are fulfilled. That should be fine, right?
