### 2019-12-20 #1969 @kbarnhart
@JesperDramsch I re-visited the [JOSS documentation section on authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship) which states:

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed. In addition, co-authors agree to be accountable for all aspects of the work, and to notify JOSS if any retraction or correction of mistakes are needed after publication.

I think you are correct to determine that authorship criteria by all three authors are fulfilled.
