### 2019-12-23 #1969 @jessepisel
Howdy @kbarnhart I am going to start my review this week. My only potential COI on this review is that I was part of the same research group just over 4 years ago. However, I have not done any research with any of the authors, and have no publications with any of them. Is it still OK for me to review the package despite the potential conflict of interest?
