### 2019-12-23 #1969 @kbarnhart
@jessepisel thanks for checking in about COI. Looking at the [JOSS COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) the most important passage is:

> As a reviewer, COIs are your present or previous association with any authors of a submission: recent (past four years) collaborators in funded research or work that is published; and lifetime for the family members, business partners, and thesis student/advisor or mentor. In addition, your recent (past year) association with the same organization of a submitter is a COI, for example, being employed at the same institution.

Based on this, my assessment is that you would have a COI if either

- you were affiliated with Mines in the last year or
- one of the authors was a thesis student, advisor, or mentor of yours

Do you meet either of these criteria? If no, then you have no COI and can proceed. If yes, then you can request that the conflict is waived (presuming you feel that you are able to make an impartial scientific judgment or evaluation).
