### 2019-12-23 #1969 @jessepisel
@kbarnhart Thank you for clarifying for me. I wanted to make sure I was interpretting the COI policy correctly. I have not been affiliated with Mines and none of the authors have been a student, advisor, or mentor. I will go ahead with my review this week.
