### 2019-12-23 rgmyr/corebreakout#5 @jessepisel
pip install error with opencv
I have been going through the install steps and ran into an error at the end when installing corebreakout in both dev mode and regular mode. From the virtual environment I run
 `pip install -e .`  or `pip install .` and get the following error for both:
`ERROR: Could not find a version that satisfies the requirement opencv (from corebreakout==0.2) (from versions: none)
ERROR: No matching distribution found for open cv (from corebreakout==0.2)`

I changed opencv to opencv-python in `setup.py` and was able to install opencv that way.

Here is some info on OS and package versions from the corebreakout virtual env.

  Date: Mon Dec 23 13:41:29 2019 CST

             Linux : OS
                 8 : CPU(s)
            x86_64 : Machine
             64bit : Architecture
            Python : Environment

  Python 3.6.9 |Anaconda, Inc.| (default, Jul 30 2019, 19:07:31)  [GCC 7.3.0]

            1.16.4 : numpy
             1.3.2 : scipy
            7.10.2 : IPython
             3.1.1 : matplotlib
             0.4.3 : scooby

  Intel(R) Math Kernel Library Version 2019.0.4 Product Build 20190411 for
  Intel(R) 64 architecture applications

Other than this error, the install went fine on this machine.

openjournals/joss-reviews#1969
