### 2019-12-31 rgmyr/corebreakout#6 @brendonhall
Brendon JOSS review items
Hi all, I've begun reviewing for openjournals/joss-reviews#1969.

I was able to install the library I think.  I've have been trying to use the library to train a model.  The documentation on how to do this is quite brief.  I think you really need an example of a typical labeling workflow, and demonstrate how to use `labelme` to do this.  Some specific items:

- what class names are needed to identify the core sticks?  I was able to see a background class, try and columns by displaying a PolygonDataset() object, but no other info about them.

- there isn't a good description of the relationship between classes and instances.  Are tray objects all one instance?  is background?  presumably core columns are not?

- how much training data is needed?

- the `train_mrcnn_model.ipynb` notebook refers to a pretrained model.  This isn't in the repo, and a didn't see a reference about where to get it

- the default directories `assets/*` don't exist

- where is the provided BGS model and data?
