### 2019-12-31 rgmyr/corebreakout#6 @zanejobe
Hi Brendon,

Thanks for taking the time to review!

I'll let Ross or Thomas answer some of the other questions, but the last three bullet points I can answer quickly. Go here https://github.com/rgmyr/corebreakout/releases and download `assets`.  There are two folders in there, one which contains the data (test and train) and another folder with two pre-trained models.
