### 2020-01-01 rgmyr/corebreakout#6 @ThomasMGeo
There are also some issues with wholesale downloading all of the BGS photo's and re-hosting them from a copyright perspective. We used 30-50 boxes, with limited uplift towards 50.
