### 2020-01-01 rgmyr/corebreakout#6 @rgmyr
Hi Brendon,

- Regarding the `assets`, we do explain where to find those in the [Getting Started section](https://github.com/rgmyr/corebreakout#download-data-optional) of the README (just above Installation). If you have suggestions about how to make that more clear/obvious, I'd be happy to update.
- I think your questions about labeled classes/instances are answered in [docs/creating_datasets.md](https://github.com/rgmyr/corebreakout/blob/master/docs/creating_datasets.md). Let me know if you've seen that... if it's not clear enough, I can add something that's more explicit about instances vs. classes. We could also add some labeled screenshots if that would help.
