### 2020-01-02 #1946 @kbarnhart
@swgeoscience no worries. I think that we are good with three reviewers. If you are interested in generally being considered as a JOSS reviewer 🎉,  please consider filling out the 1-minute form at [this link](https://joss.theoj.org/reviewer-signup.html).
