### 2020-01-06 531afb7
added missing link for training notebook
The url for the training notebook was missing, so I added it. All the other links work fine.

openjournals/joss-reviews#1969
