### 2020-01-06 f3135b4
added link to model selection notebook
added a link to the model selection notebook
