### 2020-01-06 rgmyr/corebreakout#5 @jessepisel
Here is the latest on this. After changing `opencv` to `opencv-python` in `setup.py` I was able to get the package installed and running on two different Linux machines.

Windows has a few other issues in the install with Mask RCNN (not surprised).

With this change in the setup, I have the package fully installed and am able to train on the `assets.zip` dataset.
