### 2020-01-08 7206fa4
opencv -> opencv-python in setup.py
