### 2020-01-08 rgmyr/corebreakout#5 @rgmyr
Thanks for the info, @jessepisel.

I've updated `setup.py` to list `opencv-python` instead of `opencv` in 7206fa4 and will close this issue.
