### 2020-01-24 1a851e0
add note on training images; add juptyer to reqs; delete datasets.utils
