### 2020-01-27 #1969 @kbarnhart
@jessepisel thanks for completing your review and letting me know. Once all the reviewers have completed their reviews and the authors have finished addressing them, I will ask you (and the other reviewers) to affirm recommendation to accept in JOSS.

👋 @brendonhall and @JesperDramsch a friendly reminder to complete your review. If you have a time estimate, please let me know. Thanks again for being willing to review for JOSS.
