### 2020-02-02 #1969 @JesperDramsch
I apologize for the delay. I had started my review early January, but life got in the way. I know this is frustrating, so I'll do my best to be as helpful as I can.

My main problem reviewing the package at the moment, is grasping the scope of the package and working through the documentation and API.

## Scope
The paper and description say the package is "for depth alignment and segmentation" and "corebreakout is currently utilized for ongoing work for lithology classification". Is the scope of corebreakout preparing and aligning the core images for this classification only, or also for subsequent deep learning-based lithology classification? The examples only show the extraction of the core from the full images and aligning these.

## Docs
I would not be able to recreate the examples from the available documentation. Moreover, it appears that while docstrings are available, no documentation was generated and snippets of documentation are spread throughout the structure of the project. I.e. there is a docs folder, however, utility scripts have a separate short description in the utility script folder, which however does not describe all utility scripts. Moreover, there are example notebooks in the root of the project and in the test/ directory, which is not directly clear, why these are kept separate, or how they are located separate from the other notebooks. I was not able to locate the functional API documentation, but am in contact with the authors, which makes it hard to make a comprehensive review for me.

## Improvements for now
I do see the value of the project, and would very much like to see it get accepted. I just have trouble grasping the full scope and navigating the code at this point, which I think should be addressed to make this project as useful as possible to the scientific and FOSS community. Happy to help where I can.

Does it make sense to open this as an issue on the project itself?
