### 2020-02-03 #1969 @kbarnhart
@JesperDramsch thanks for the comments here and for your work reviewing this contribution. I would recommend that you create at least two (maybe three) issues in the main project repository. If you could tag this issue (openjournals/joss-reviews/issues/1969) it will help create a link between the changes in the main repository and this review.

I would make a separate issue for each major concern you have. That way the authors can address each individually, and you can weigh in on whether the issue was fixed. Based on your comment above, it seems like the main issues you identify are:
* the scope is not clear.
* you were not able to recreate the examples from the documentation.
* you have some concern regarding the organization of the documentation and example notebooks

Let me know if you have any questions.
