### 2020-02-04 rgmyr/corebreakout#10 @JesperDramsch
Question about Scope and paper.md [JOSS Review]
(Apologize for being so late, I had some unforeseen time-critical things come up with my thesis. I'll do my best to help! I apologize if I misunderstand some things here, English is only my second language, so I hope that was not the main barrier here. Also, I'm not trying to be overly critical and hope this can be read as constructive. I'm happy to help make this as good as possible. I definitely see the value in this project.)

I have read the paper and the description in this repo, and made an initial comment in [Joss Review 1969](https://github.com/openjournals/joss-reviews/issues/1969)

The paper summary gives an overview of core, however only reserves a single paragraph to the actual project, and a single sentence to the scope of the project. The description on the project readme.md:

> corebreakout is a Python package built around matterport/Mask_RCNN for the segmentation and depth-alignment of geological core sample images.
> It also provides a CoreColumn data structure, which can be sliced and added, and which facilitates saving, loading, manipulating, and visualizing depth-aligned core image data.
> We are currently using this package to enable research on Lithology Prediction of Slabbed Core Photos Using Machine Learning Models, and are working on getting a DOI for the project through the Journal of Open Source Software.

This description is vague enough in the sense, that it is unclear to me what either of these parts do. The paper.md suffers from the same points, I believe, which is in essence, that you as authors of the package know exactly what it does, but it's not directly clear to new people coming in.

So let me summarize my issues:

## What is the actual scope of the project?
- Are you taking raw box photos, extracting the ROI with the RCNN of the core in those photos, then stitching them?
- Are you also doing lithological segmentation with this, as implied by the overview and the paper.md? If yes, that would be a nice example notebook or documented somewhere

## What do you train on?
In the paper you say you make it straightforward for geologists to add their own labeled images. I believe this is also mentioned in #6 . In the documentation, I have only seen a reference to labelme. It is not clear what actually has to be labeled, and how the model generalizes. Do you just draw rectangles around the core? What classes are needed? How do you actually label the ruler?

It's not clear how it comes together. I understand that you should not reproduce the entire labelme documentation. But a minimally viable tutorial would be really helpful for anyone trying to use the software.

## Does it generalize?
You use deep learning. Is it enough to train on a few samples and get good results on all the other unlabeled fotos? Do you have an idea how many are enough to achieve reliable results?

## Paper functionality
In the functionality section you describe quickly what you can do and how users can use it.

However, you mention a possible GUI that is not provided and work that is using corebreakout to do lith classification (hence my further confusion). This is the first mention of the CoreColumn class, which seems to be at the core of the package.

I believe going more into the details of what CoreColumn actually does, would be extremely benefitial. Like I mentioned, you know exactly what it does, from the descriptions I read here, I only know that it provides a convenient interface for retrieving, manipulatin, visualizing, and slicing data.

## OCR
You mention you provide all the codes for OCR. I think it would be benefitial to mention that additionally Tesseract has to be installed for this. I was going through the setup, when I happened upon the OCR section, which told me to install even more, which is a bit unfortunate.
