### 2020-02-04 rgmyr/corebreakout#11 @JesperDramsch
Documentation Improvements [JOSS Review]
I would not be able to recreate the examples from the available documentation.

It appears that while docstrings are available, no documentation was generated and snippets of documentation are spread throughout the structure of the project. I.e. there is a [docs folder](https://github.com/rgmyr/corebreakout/tree/master/docs), however, utility scripts have a separate short description in the [utility script folder](https://github.com/rgmyr/corebreakout/tree/master/scripts/post_labeling_tools), which however does not describe all utility scripts. Moreover, there are [example notebooks](https://github.com/rgmyr/corebreakout/tree/master/notebooks) in the root of the project and in the [test/ directory](https://github.com/rgmyr/corebreakout/tree/master/tests/notebooks), which is not directly clear, why these are kept separate, or how they are located separately from the other notebooks.

## Functional API
I was not able to locate the functional API documentation. I see docstrings are available, but I do not see any documentation of the functions and classes.

From reading the readme.md and the paper I am guessing that there is a `CoreColumn`, which I don't see any description of other than "it is essential to the project". And it seems there is a `CoreSegmenter` which apparently has a `.segment` and a `segment_all` method. Seeing that you have a docstring, what was the choice to not auto-generate API documentation that people can use?

I don't think it's good for the project adoption if users have to sift through the code to find the documentation (I remember in one document, you do link directly to the code "for more information" but I can't find it right now). I don't think this is good practice, and creating central documentation to get information about usage, without having to deeply understand the design choices you the authors make about placing classes, functions and methods, would be extremely beneficial and quite frankly necessary.
