### 2020-02-04 rgmyr/corebreakout#11 @JesperDramsch
[Link to Issue](https://github.com/openjournals/joss-reviews/issues/1969)
