### 2020-02-04 rgmyr/corebreakout#10 @ThomasMGeo
> What is the actual scope of the project?

> Are you taking raw box photos, extracting the ROI with the RCNN of the core in those photos, then stitching them?

Yes!

> Are you also doing lithological segmentation with this, as implied by the overview and the paper.md? If yes, that would be a nice example notebook or documented somewhere

The core columns have been used for this type of work, but this is out of scope for this paper.
