### 2020-02-04 rgmyr/corebreakout#10 @ThomasMGeo
> Does it generalize?
> You use deep learning. Is it enough to train on a few samples and get good results on all the other unlabeled fotos? Do you have an idea how many are enough to achieve reliable results?

In our research, around ~30 boxes you start to get limited improvement after the fact. This was only done with the BGS / UK style tray's. Unsure how many you would need for other trays. Some of the core data providers provide really nice trays (almost all completely pure black) while Norway is all over the place.
