### 2020-02-06 rgmyr/corebreakout#10 @JesperDramsch
I think that information would be extremely useful in the documentation for new users getting started and clarify. Thank you for adding these points here.
