### 2020-02-10 #1969 @JesperDramsch
I can proceed with the functional review once the documentation issues are addressed. I don't have time testing the full functionality by sifting through raw do strings I'm afraid.
