### 2020-02-10 #1969 @kbarnhart
@JesperDramsch thanks for the comment. It is, of course, reasonable to for you to wait until the issues you identified in rgmyr/corebreakout#11 be addressed before you proceed with the review.

Thank you again for contributing to this review.
