### 2020-02-17 rgmyr/corebreakout#11 @rgmyr
Hi Jesper,

Thanks for your feedback. I'm certainly open to reorganizing the docs to be more useful/helpful. To answer a few questions:
- I didn't auto-generate API docs just because I've had issues getting that stuff (e.g., `sphinx`) to work well in the past. I'll look into it again and give it a shot though. I'll also try to clean up the scripts folder + their docs.
- The notebooks at `tests/notebooks` are separate because they're meant to be run as 'tests' (i.e., after making changes to make sure that column plots / segmentation outputs are still as expected). And `make_endpts_data.ipynb` was just used to generate 'correct' answers for some `pytest` tests and make some doc images.
