### 2020-02-18 rgmyr/corebreakout#11 @JesperDramsch
Yeah, I've heard that about sphinx before. However, as it stands I don't think the documentation is sufficient for me to figure out how to use the library.
