### 2020-02-22 #1969 @kbarnhart
@JesperDramsch thanks for letting me know. I won't expect to hear from you until March 20.

JOSS has [an out of office feature](https://joss.readthedocs.io/en/latest/editing.html#out-of-office
) (I know that it works for editors, am not 100% that it works for reviewers, but I expect it might).

If you write a comment like:

Then anytime someone mentions you in this or another issue in this repository the bot will respond indicating that you are ooo until the end date. Feel free to try and use this if you think it would be helpful.
