### 2020-03-09 #1969 @kbarnhart
@rgmyr I wanted to check in and get a quick update on where you stand on addressing the current set of outstanding issues.
