### 2020-04-17 2d73cf9
Paper and README changes on scope/functionality/reqs; first round of sphinx-generated html docs
