+++ docs/index.rst
@@ -1,0 +1,24 @@
+.. CoreBreakout documentation master file, created by
+   sphinx-quickstart on Fri Apr 17 06:21:33 2020.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to CoreBreakout's documentation!
+========================================
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Documentation Contents:
+
+   creating_datasets
+   model_building
+   layout_parameters
+
+Indices and tables
+==================
+
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
