### 2020-04-17 rgmyr/corebreakout#11 @rgmyr
Hi @JesperDramsch, I've finally gotten around to this and have compiled some html documentation pages. I am still cleaning up some of the `.rst` structure, and plan to host them on a `readthedocs` page in the near future, but for now you can view them with a simple HTTP server (e.g., `python -m SimpleHTTPServer`) in the `docs/_build` folder if you are interested.
