### 2020-04-17 rgmyr/corebreakout#10 @rgmyr
@JesperDramsch I've updated the README and paper.md to hopefully address these concerns to some extent. In particular, I've tried to make the scope more clear the in the paper and README intro, I've added a note about training images in two separate places, and I've added an `Additional (Optional) Requirements` section to the README to mention the OCR more up front.

I also plan to add a more succinct overview section to the sphinx html docs, and possibly a CoreColumn tutorial (which will probably be a modified version of the test notebook). I can add some detail to the dataset creation guidelines, but I'm not sure how much more detailed that can be in words, so I would appreciate any concrete suggestions of what would make that section more clear or easy to follow. Thanks!
