### 2020-04-17 #1969 @rgmyr
Hi all, apologies for the wait. I've made some changes and additions to at least partially address both of @JesperDramsch's open issues. I expect there to be a few more changes before he is satisfied, but I also expect to be able to work on this much more in the short term in response to his feedback.
