### 2020-04-22 7c3df41
docs and formatting updates
