+++ requirements.txt
@@ -7,7 +7,7 @@ cython
 matplotlib
 scikit-image
 keras>=2.0.8,<=2.2.5
-opencv
+opencv-python
 h5py
 jupyter
 IPython[all]

[…]
