+++ .readthedocs.yml
@@ -16,6 +16,9 @@ sphinx:
 # Optionally build your docs in additional formats such as PDF and ePub
 formats: all
 
+submodules:
+  include: all
+
 # Optionally set the version of Python and requirements required to build your docs
 python:
   version: 3.7
