+++ .readthedocs.yml
@@ -21,10 +21,8 @@ submodules:
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-  version: 3.7
+  version: 3.6
   install:
-    - requirements: requirements.txt
+    - requirements: docs/rtd_requirements.txt
     - method: pip
       path: Mask_RCNN
-      extra_requirements:
-        - tensorflow
