### 2020-05-03 rgmyr/corebreakout#10 @JesperDramsch
I'll check them out thank you!

It's hard for me to say, as I have no documentation to understand the code or how to use it, to make those suggestions. I think Brendon's suggestions in #6 are a good start.

Generally, it should be clear, what CoreColumn is, does, how to instantiate it and how to then use it, how it interacts or binds into the rest of the library. I think...
