### 2020-05-05 5a4641c
fill out doc pages; add CoreColumn demo notebook + scripts reference page
