+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Built `sphinx`-generated html docs
 - Implemented `CoreColumn.iter_chunks()`
-
+- Added a `CoreColumn` demo notebook
 
 ## 0.2
 

[…]
