### 2020-05-05 rgmyr/corebreakout#10 @rgmyr
Hi @JesperDramsch, the docs page has been filled out, and I'm hoping that things are clear enough to follow at this point. I've also added a `column_demo.ipynb` notebook, which is linked to in the docs.

Please let me know if you get stuck anywhere or there is something specific that is lacking. Thanks!
