### 2020-05-06 rgmyr/corebreakout#6 @ThomasMGeo
Hi @brendonhall , the documentation just was updated, give it another look if you have a chance.
