### 2020-05-08 #1969 @kbarnhart
@rgmyr @ThomasMGeo the new hosted docs https://corebreakout.readthedocs.io/en/latest/ look great!

I wanted to check in and see whether reviewers @JesperDramsch and @brendonhall should tune back in (per their ability to do so), or whether you are still working to address the issues they raised.
