### 2020-05-18 #1969 @brendonhall
Hi everyone.  Glad to see the progress on this. I was able to follow the installation instructions, and get everything installed and the dev environment set up.  I'm quite familiar with the workflow this library is trying to enable, and have written many scripts/ML models to extract and analyze core images.  So I'm quite excited about this library.

Some things I've noticed so far:
- requirements doesn't include tensorflow.  needed to install this to run the notebooks.
- I've been running the notebooks 'out of the box'.  I noticed a couple of errors:
- I installed the `assets.zip` file as instructed.  I get the following error when running the last cell of `select_model.ipynb`

```OSError: Unable to open file (unable to open file: name = '/home/brendon/Projects/corebreakout/assets/models/retrain/cb_default20191104T1443/mask_rcnn_cb_default_0030.h5', errno = 2, error message = 'No such file or directory', flags = 0, o_flags = 0)```

and this error when running the line `model.load_weights(str(model_loc / 'mask_rcnn_cb_default_0100.h5'), by_name=True)` from `train_mrcnn_model.ipynb`:
```OSError: Unable to open file (unable to open file: name = '/home/brendon/Projects/corebreakout/assets/models/retrain/cb_default20190930T1243/mask_rcnn_cb_default_0100.h5', errno = 2, error message = 'No such file or directory', flags = 0, o_flags = 0)```
