### 2020-05-18 #1969 @brendonhall
I'm still finding to hard to figure out how to 'get started'.  It would be great if there was a clear tutorial, ideally using a jupyter notebook.  It should run `out of the box` as much as possible, using the provided data.  Besides running the notebooks that are supplied (and see errors above), the workflow to perform a classification isn't clear.

Also, the description of how to label the cores isn't clear.  The docs state
> At a minimum, you will want some class to represent columns. You may also create a class (e.g., 'tray')...
This is be more specific.  'some class' leaves questions.  Do we need to label core/not core, core, tray, etc?  Do labels need to be dense or sparse?  Some diagrams here and a best practice would be very useful.  Are there labels supplied in assets?  Can you point that out?
