### 2020-05-21 rgmyr/corebreakout#14 @rgmyr
Brendon Review Item: tensorflow requirement
Hi @brendonhall, I'm just making a separate thread for this particular issue.

As noted in the README instructions:

> The TensorFlow requirement is not explicitly listed in `requirements.txt` due to the ambiguity between `tensorflow` and `tensorflow-gpu` in versions `<=1.14`...

Also the installation instructions include installation of `tensorflow-gpu` and a note about the CPU-only option, before installing `corebreakout`:
```
$ conda create -n corebreakout-env python=3.6 tensorflow-gpu=1.14
$ conda activate corebreakout-env
```

>**Note:** If you want to try a CPU-only installation, then replace `tensorflow-gpu` with `tensorflow`. You may also lower the version number if you are on a machine with `CUDA<10.0` (required for TensorFlow`>=1.13`). See [TensorFlow GPU requirements](https://www.tensorflow.org/install/gpu#software_requirements) for more compatibility details.

This is a known issue with `tensorflow` and PyPI (See [tensorflow Issue #7166](https://github.com/tensorflow/tensorflow/issues/7166)), and there's no great workaround that I'm aware of, at least for versions `<1.15`.

In 96fbcae I've added a `UserWarning` mechanism to the `setup.py` so that users won't be able to install `corebreakout` until they have manually installed `tensorflow` or `tensorflow-gpu`.

If you know of some preferable workaround, I'm happy to implement it. Otherwise, I'm not sure how else to remedy that.
