### 2020-05-24 12e27c1
add inspect_data.ipynb + expand docs on datasets/labeling
