+++ setup.py
@@ -10,7 +10,7 @@ except ImportError:
 try:
     import tensorflow
 except ImportError:
-    raise UserWarning('tensorflow or tensorflow-gpu must be installed manually!')
+    raise UserWarning('`tensorflow` or `tensorflow-gpu` must be installed manually!')
 
 
 PACKAGE_PATH = os.path.abspath(os.path.join(__file__, os.pardir))
