+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Built `sphinx`-generated html docs
 - Implemented `CoreColumn.iter_chunks()`
 - Added a `CoreColumn` demo notebook
+- Added restriction in `setup.py` to require `tensorflow` before install
 
 ## 0.2
 
