+++ docs/creating_datasets.rst
@@ -3,41 +3,89 @@
 Creating ``Datasets``
 =============================
 
-Image Labeling Guidelines
-----------------------------------------
+Using ``labelme``
+-----------------
 
-The recommended way to add a new set of labeled training images is to
-annotate them using
-`wkentaro/labelme <https://github.com/wkentaro/labelme>`__. The
-``labelme`` GUI allows the user to draw any number of labeled polygons
-on an image, and saves the labels and coordinates in a JSON annotation
-file.
+The recommended way to add a new set of labeled training images is to annotate them using `wkentaro/labelme <https://github.com/wkentaro/labelme>`__. The ``labelme`` GUI allows the user to draw any number of labeled polygons on an image, and saves the labels and coordinates in a JSON annotation file.
 
-Simply copy all the images you want to label into a flat directory, open the directory in ``labelme``, and begin saving your annotations. To be able to use the built-in ``corebreakout.datasets.PolygonDataset`` class with your training data, you will want to follow these labeling guidelines:
+You can easily install ``labelme`` with ``pip``, and then open our small test dataset:
+
+.. code::
+
+  $ pip install labelme
+  $ labelme tests/data/two_image_dataset
+
+This should open a window that looks like this:
+
+|image0|
+
+New polygons can be drawn by clicking ``Create Polygons`` and then clicking points on the image to add vertices. When get back around to the starting vertex to close the loop of points and create a polygon, ``labelme`` will ask you to assign it a text label.
+
+Unique text labels are aggregated on the right side, and the individual polygons shown in the image are colored and listed with their assigned labels. They happen to be the same in this example, but there could be more labels in the ``Label List`` than polygons in the ``Polygon Labels`` list for any given image in the directory.
+
+When you finish adding polygons, click the ``Save`` button to export the annotations as a ``.json`` file, which by default will save to the current directory.
+
+The ``File List`` shows images in the currently directory, with the checkbox indicating whether an annotation has been saved for that file yet. You can use the ``Next Image`` and ``Previous Image`` buttons to traverse through the images (and corresponding annotations) in the directory.
+
+
+Inspect Our Annotations
+-----------------------
+
+For a more hands-on exploration of annotation structure and the ``Dataset`` API, see `notebooks/inspect_dataset.py <https://github.com/rgmyr/corebreakout/blob/master/notebooks/inspect_dataset.ipynb>`__
+
+
+How to Name Things
+------------------
+
+As you can see above, our text labels for individual objects have the form ``<class><number>``. The ``class`` refers to a general object type, the ``number`` differentiates between individual `instances` of that class within an image. This isn't the only possible naming scheme, but it's the one you should use if you want your annotations to be compatible with our ``PolygonDataset`` implementation.
+
+In our data, we use two classes, **col** and **tray**, to label two different types of objects:
+
+- **col** is for `columns` of core material -- the individual objects that we want to segment, crop, and stack on top of each other.
+- **tray** is for empty trays, which for our data are the most reliable "measuring stick" to use for defining the "top" and "base" positions of columns within an image
+
+The first type of object is required -- the whole purpose of this workflow is to segment core columns, so you will need to at least label those.
+
+The second type of object is optional -- using another object (empty tray, measuring stick, etc.) to define top and base positions of columns is `one` option for the ``endpts`` setting in :ref:`layout-parameters`, but there are others which do not require the existence of any objects other than core columns.
+
[…]
