### 2020-05-24 rgmyr/corebreakout#6 @rgmyr
Hi @brendonhall, I'm just going to use this thread to address some of your other recent comments.

I'm going to try and pull together something that's a bit more out-of-the-box, but still with good coverage of the functionality. It's a little harder to do that with `select_model.ipynb` in particular, because that notebook presupposes that you have a folder full of model checkpoints, which I'm reluctant to provide since it would increase the size of `assets.zip` by a lot. At the very least I'll add some markdown to the notebook to do a better job of explaining the usage -- it should be fairly usable *after* you've done a training run.

In any case, I have added a new notebook called `inspect_dataset.ipynb`, which, as you might expect, inspects the small dataset in `tests/data` (which has all the same general properties as the datasets in `assets`). This is tied together with an expansion of the documentation on [Creating Datasets](https://corebreakout.readthedocs.io/en/latest/creating_datasets.html). I may yet flesh that out or reorganize a bit more, but at the very least I hope it's more clear about how to use `labelme`, and the distinction between classes/instances.
