### 2020-05-24 #1969 @rgmyr
Hi all -- just to give an update:

I've started addressing Brendon's issues on the submission repo [here](https://github.com/rgmyr/corebreakout/issues/14) and [here](https://github.com/rgmyr/corebreakout/issues/6).
