### 2020-05-26 rgmyr/corebreakout#14 @kbarnhart
linking to joss review openjournals/joss-reviews/issues/1969
