### 2020-06-10 rgmyr/corebreakout#6 @rgmyr
I've just pushed another commit which adds some markdown instructions and comments to `select_model.ipynb`.

I've also added a new notebook called `segmenter_demo.ipynb`, which demos the segmentation API and should work "out of the box".
