### 2020-06-10 #1969 @rgmyr
Hi @kbarnhart, I pushed another update this morning, which is meant to address Brendon's issues with the `select_model` notebook by adding some markdown instructions/comments/etc., and adds a new `segmenter_demo` notebook (to complement the `column_demo` notebook) in response to his request for something more "out of the box".

We probably won't add anything else before Brendon and/or Jesper are able to revisit their reviews.
