### 2020-06-11 #1969 @kbarnhart
@rgmyr thanks for the update. I appreciate it and all the updates you guys have made. They look great.

@brendonhall and @JesperDramsch it looks like this submission is now ready for you guys to look at again. I believe that the authors have revised in response to your initial comments such that they expect you will be able to complete your reviews. Please let me know when you anticipate being able to complete your reviews. As always, if you have questions, please feel free to reach out to me.

Thanks again to all for your participation in this iterative review process.
