### 2020-06-15 #1969 @kbarnhart
:wave: @brendonhall @JesperDramsch a quick ping to see if you could revisit your reviews and (potentially) complete the JOSS review checklist.
