### 2020-06-18 #1969 @kbarnhart
👋 @brendonhall @JesperDramsch another quick ping to see if you could revisit your reviews and (potentially) complete the JOSS review checklist.
