### 2020-06-18 rgmyr/corebreakout#6 @brendonhall
Hi @rgmyr, this thread is fine and thanks for getting back to me.  And yeah, my comments a while ago were similar.  When I start using a new library, I basically look for how to get started as quickly as possible.  Usually a tutorial, that I can get working and then start modifying that for my specific problem.

`inspect_dataset.ipynb` is very useful in this regard.  For example pointing out that the background class is implicitly assigned.  This does mean though that you must label ALL non-background objects in an image, correct?

I appreciate you attention to the comments that I've made.  This is a useful tool for researchers as well as having industrial applicaitons (energy, mining)
