### 2020-06-18 #1969 @brendonhall
Hi @kbarnhart, @rgmyr, I have completed my review of corebreakout.  The authors have done a great job of addressing the questions I had, and I feel this is ready for publication. The documentation and notebooks provide the background necessary to use the library.  I feel that this is an important contribution to digital subsurface science, and I plan on applying this library to some of my own work as well.

Please get in touch if you have any questions or if I can help at all.
