### 2020-06-18 #1969 @brendonhall
no, sorry that should have been checked off.  I'm actually not aware of any other published works that deal with this extracting core images like this.  There was the Get 2 the Core challenge a couple of years ago (https://www.youtube.com/watch?v=N8CvQUE3L0c) but I didn't hear anything about that after it started.
