### 2020-06-18 #1969 @JesperDramsch
Is it possible that someone accidentally checked off my checklist?

I was just finishing up my review of the software and always made sure not to check off anything I hadn't done yet, but just returned to a full checklist...

Nevertheless, I have also completed my work and think the library is at an acceptable state with some really cool work I'm happy the authors have published. Thanks for the additional work on this, I think the documentation really has made huge strides and helped me understand the utility.
