### 2020-06-18 #1969 @kbarnhart
@JesperDramsch thanks for the update. I'm not sure how to see a record of the checklist (or if its saved...) so I can't tell when/how your checklist was edited. It looked incomplete to me this morning when I pinged.

Either way, it sounds like you think the software merits a full checklist from you (please let me know if I'm mistaken).
