### 2020-06-18 #1969 @kbarnhart
@rgmyr and co-authors I'll be able to do my final checks by the end of tomorrow and then expect to move this forward to the final steps of the JOSS process. At most you may see a request from me with some paper copy editing or a PR with these edits.

Thanks to all three of the reviewers @jessepisel @brendonhall and @JesperDramsch for providing feedback and working with the authors over multiple iterations. I appreciate your contributions to the JOSS review process.
