### 2020-06-18 6203528
Update paper.bib
Added a comma to line 94, getting a compile error on line 89.
