+++ paper.bib
@@ -91,6 +91,6 @@
   	Booktitle = {The Geological Interpretation of Well Logs: Third Edition, by Malcom Rider and Martin Kennedy.~ISBN 978-09541906-8-2 (PB).~Published by Rider-French Consulting Limited, Glasgow, Scotland, 2011.},
   	Publisher = {Rider-French Consulting Limited},
   	Title = {{The Geological Interpretation of Well Logs: Third Edition}},
-    ISBN = 978-09541906-8-2
+    ISBN = 978-09541906-8-2,
   	Year = 2011
 }
