### 2020-06-18 bb13084
Update paper.bib
Is this the proper way to do a bibtex with two authors? It should be Rider and Kennedy in the text.
