+++ paper.bib
@@ -86,7 +86,7 @@
   address = {Washington, DC, USA},
 }
 
-@book{Rider and Kennedy:2011,
+@book{{Rider and Kennedy}:2011,
   	Author = {{Rider}, M. and {Kennedy}, M.},
   	Booktitle = {The Geological Interpretation of Well Logs: Third Edition, by Malcom Rider and Martin Kennedy.~ISBN 978-09541906-8-2 (PB).~Published by Rider-French Consulting Limited, Glasgow, Scotland, 2011.},
   	Publisher = {Rider-French Consulting Limited},
