### 2020-06-18 #1969 @kbarnhart
@rgmyr there are four issues in the CoreBreakout repository that are associated with this review and still open. If they have been addressed, will you close them?
