### 2020-06-18 rgmyr/corebreakout#6 @rgmyr
> This does mean though that you must label ALL non-background objects in an image, correct?

Yes, but only the objects that need to be detected because they have some purpose in processing (either core columns or "measuring stick" type objects that you intend to actually use as such). The background doesn't need to be visually uniform, and is simply a catch-all for everything that's not directly useful.

I'm going to close this issue so that Katy can go ahead with finalizing the review, but feel free to open another issue if you use this project in the future and run into any problems.

Thanks for your time and constructive feedback!
