### 2020-06-18 #1969 @kbarnhart
@rgmyr at this point could you:
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository)
- [ ] Ensure that the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (should match the author list and not include people who only made a small fix); you may also add the authors' ORCID.
