### 2020-06-18 d0339b5
update CHANGELOG, fix link to v0.2 release.
