### 2020-06-19 #1969 @rgmyr
Hi @kbarnhart, thanks for the help and instructions.

- Tagged release is [v0.3](https://github.com/rgmyr/corebreakout/releases/tag/v0.3)
- [Link to corresponding repo tag](https://github.com/rgmyr/corebreakout/tree/v0.3)
- Zenodo DOI: `10.5281/zenodo.3900640`
- [Link to Zenodo archive](https://zenodo.org/record/3900640)

Thank you to @brendonhall @JesperDramsch and @jessepisel for your time and constructive feedback!
