### 2020-06-19 #1969 @kbarnhart
@rgmyr I've now recommended to the JOSS editors on chief that this paper be accepted and published.
