### 2020-06-19 #1969 @ThomasMGeo
The formatting is a little goofy on that paper.bib entry. Happy to dig into it @kbarnhart.
