### 2020-06-19 d9fcf2f
Update paper.bib
Looking into Rider French.
