+++ paper.bib
@@ -88,7 +88,6 @@
 
 @book{Rider:2011,
   	Author = {{Rider}, M. and {Kennedy}, M.},
-  	Booktitle = {The Geological Interpretation of Well Logs: Third Edition, by Malcom Rider and Martin Kennedy},
   	Publisher = {Rider-French Consulting Limited},
   	Title = {{The Geological Interpretation of Well Logs: Third Edition}},
     ISBN = 978-09541906-8-2,
