+++ paper.bib
@@ -87,9 +87,9 @@
 }
 
 @book{Rider:2011,
-  	Author = {{Rider}, M. and {Kennedy}, M.},
-  	Publisher = {Rider-French Consulting Limited},
-  	Title = {{The Geological Interpretation of Well Logs: Third Edition}},
-    ISBN = 978-09541906-8-2,
-  	Year = 2011
+  Author = {{Rider}, M. and {Kennedy}, M.},
+  Publisher = {Rider-French Consulting Limited},
+  Title = {{The Geological Interpretation of Well Logs: Third Edition}},
+  ISBN = {978-09541906-8-2},
+  Year = {2011}
 }
