### 2020-06-19 #1969 @ThomasMGeo
I just cleaned up that entry, if it still fails, unsure of next steps. It had some duplicate information in the paper.bib and added some more standard formatting.
