### 2020-06-19 #1969 @danielskatz
It looks like those changes fixed the problem.
