### 2020-06-19 #1969 @danielskatz
I'm going to suggest some changes for style in a few minutes
