### 2020-06-19 e444dad
minor style changes
