+++ paper.bib
@@ -3,7 +3,7 @@
                Georgia Gkioxari and
                Piotr Doll{\'{a}}r and
                Ross B. Girshick},
-  title     = {Mask R-CNN},
+  title     = {Mask {R-CNN}},
   journal   = {CoRR},
   volume    = {abs/1703.06870},
   year      = {2017},
@@ -46,7 +46,7 @@
           Minisini, Daniel and
           Howes, Nick and
           Cantelli, Alessandro},
-    title = {Facies architecture of submarine channel deposits on the western Niger Delta slope: Implications for grain-size and density stratification in turbidity currents},
+    title = {Facies architecture of submarine channel deposits on the western {N}iger {D}elta slope: Implications for grain-size and density stratification in turbidity currents},
     journal = {Journal of Geophysical Research: Earth Surface},
     volume = {122},
     number = {2},
@@ -59,10 +59,9 @@
 }
 
 @misc{Abdulla:2017,
-  title={Mask R-CNN for object detection and instance segmentation on Keras and TensorFlow},
+  title={Mask {R-CNN} for object detection and instance segmentation on {K}eras and {TensorFlow}},
   author={Waleed Abdulla},
   year={2017},
-  publisher={Github},
   journal={GitHub repository},
   howpublished={\url{https://github.com/matterport/Mask_RCNN}},
 }
@@ -71,13 +70,14 @@
   author =       {Ketaro Wada},
   title =        {{labelme: Image Polygonal Annotation with Python}},
   howpublished = {\url{https://github.com/wkentaro/labelme}},
+  journal={GitHub repository},
   year =         {2016}
 }
 
 @inproceedings{Smith:2007,
   author = {Ray Smith},
-  title = {An Overview of the Tesseract OCR Engine},
-  booktitle = {ICDAR '07: Proceedings of the Ninth International Conference on Document Analysis and Recognition},
+  title = {An Overview of the {T}esseract {OCR} Engine},
+  booktitle = {{ICDAR '07: Proceedings of the Ninth International Conference on Document Analysis and Recognition}},
   url = {http://www.google.de/research/pubs/archive/33418.pdf},
   year = {2007},
   isbn = {0-7695-2822-8},
