### 2020-06-19 #1969 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
